requirements.txt
torch
tensorflow>=2.1.0
transformers
uvicorn
aiofiles
fastapi
elasticsearch>=7.7.1
spacy
numpy
torchvision
# DeSOTA Pkcgs
requests
pyyaml